+++ playground/example-keycloak-private-key-jwt.R
@@ -53,6 +53,7 @@ library(shiny)
 
 # ---------- Config ----------
 issuer <- "http://localhost:8080/realms/master"
+options(shinyOAuth.allow_insecure_oidc_loopback = TRUE)
 
 client_id <- Sys.getenv("OAUTH_CLIENT_ID", unset = "so-pk")
 key_path <- Sys.getenv("OAUTH_CLIENT_PRIVATE_KEY_PATH", unset = "")